requirements.txt
streamlit
pandas
pycountry
google-generativeai>=0.4.0
python-dotenv
